move FaraoImportException to api package
