+++ data/crac-creation/crac-creator-api/src/main/java/com/farao_community/farao/data/crac_creation/creator/api/FaraoImportException.java
@@ -5,9 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package com.farao_community.farao.data.crac_creation.creator.cim.crac_creator;
-
-import com.farao_community.farao.data.crac_creation.creator.api.ImportStatus;
+package com.farao_community.farao.data.crac_creation.creator.api;
 
 /**
  * @author Peter Mitri {@literal <peter.mitri at rte-france.com>}
