--- data/crac-creation/crac-creator-csa-profiles/src/main/java/com/farao_community/farao/data/crac_creation/creator/csa_profile/crac_creator/CsaProfilesImportException.java
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2023, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.farao_community.farao.data.crac_creation.creator.csa_profile.crac_creator;
-
-/**
- * @author Mohamed Ben-rejeb {@literal <mohamed.ben-rejeb at rte-france.com>}
- */
-public class CsaProfilesImportException extends RuntimeException {
-    public CsaProfilesImportException(String message) {
-        super(message);
-    }
-
-    public CsaProfilesImportException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
-
